Separate profile and user model
