+++ core/accounts/models/__init__.py
@@ -1,0 +1,2 @@
+from .users import *
+from .profiles import *
