+++ core/accounts/models/users.py
@@ -1,19 +1,17 @@
 from django.db import models
-from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin, BaseUserManager
+from django.contrib.auth.models import (
+    BaseUserManager, AbstractBaseUser, PermissionsMixin)
 from django.utils.translation import gettext_lazy as _
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 
 # Create your models here.
-
 class UserManager(BaseUserManager):
     """
-    User model manager where email is the unique identifiers
+    Custom user model manager where email is the unique identifiers
     for authentication instead of usernames.
     """
     def create_user(self, email, password, **extra_fields):
         """
-        Create and save a user with the given email and password.
+        Create and save a user with the given email and password and extra data.
         """
         if not email:
             raise ValueError(_("The Email must be set"))
@@ -22,7 +20,6 @@ class UserManager(BaseUserManager):
         user.set_password(password)
         user.save()
         return user
-
     def create_superuser(self, email, password, **extra_fields):
         """
         Create and save a SuperUser with the given email and password.
@@ -30,50 +27,34 @@ class UserManager(BaseUserManager):
         extra_fields.setdefault("is_staff", True)
         extra_fields.setdefault("is_superuser", True)
         extra_fields.setdefault("is_active", True)
-
+        extra_fields.setdefault("is_verified", True)
+        
         if extra_fields.get("is_staff") is not True:
             raise ValueError(_("Superuser must have is_staff=True."))
         if extra_fields.get("is_superuser") is not True:
             raise ValueError(_("Superuser must have is_superuser=True."))
+        
         return self.create_user(email, password, **extra_fields)
+        
 
 
 class User(AbstractBaseUser, PermissionsMixin):
-    """
-    Custom user model for our app.
+    """ 
+    Custom User Model for our app.
     """
     email = models.EmailField(max_length=255, unique=True)
     is_superuser = models.BooleanField(default=False)
     is_staff = models.BooleanField(default=False)
     is_active = models.BooleanField(default=True)
-    # is_verified = models.BooleanField(default=False)
+    is_verified = models.BooleanField(default=False)
     
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS = []
     
-    created_date = models.DateTimeField(auto_now_add=True)
-    updated_date = models.DateTimeField(auto_now=True)
-    
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
     objects = UserManager()
+    
     def __str__(self):
         return self.email
-    
-    
-class Profile(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    first_name = models.CharField(max_length=250)
-    last_name = models.CharField(max_length=250)
-    image = models.ImageField(null=True, blank=True)
-    description = models.TextField()
-    created_date = models.DateTimeField(auto_now_add=True)
-    updated_date = models.DateTimeField(auto_now=True)
-    
-    def __str__(self):
-        return self.user.email
-    
-
-@receiver(post_save, sender=User)
-def save_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
